pyproject.toml
[project]
name = "genjaxmix"
version = "0.1.0"
description = "dpmm in genjax"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "genjax",
    "graphviz>=0.20.3",
    "jaxtyping>=0.2.36",
    "plum-dispatch>=2.5.7",
]

[project.optional-dependencies]
m1 = [
    "jax-metal>=0.1.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocstrings-python>=1.14.2",
    "mkdocstrings>=0.28.0",
    "mkdocs-jupyter>=0.25.1",
    "jupytext>=1.16.6",
]
